+++ src/services/moderation/ai/handleContentModerationIgnore.service.ts
@@ -1,13 +1,13 @@
-import { makeJobId } from "../../../utils/job/makeJobId.util.js";
-import { queueContentPipelineRoute } from "../../../utils/question/pipelineRouting.util.js";
+import type { LoadedModerationContent } from "./loadModerationContent.service.js";
+import { type ModeratableContentType } from "./contentModeration.shared.js";
 
 import applyContentModerationDecisionService from "../applyContentModerationDecision.service.js";
-import { type ModeratableContentType } from "./contentModeration.shared.js";
+import { queueContentPipelineRoute } from "../../question/pipelineRouter/pipelineRouting.service.js";
+
+import { makeJobId } from "../../../utils/job/makeJobId.util.js";
 
 import moderationMetricsQueue from "../../../queues/moderationMetrics.queue.js";
 import moderationAuditQueue from "../../../queues/moderationAudit.queue.js";
-
-import type { LoadedModerationContent } from "./loadModerationContent.service.js";
 
 type HandleContentModerationIgnoreInput = {
   contentId: string;
