+++ src/services/question/ai/aiAnswer/createFeedbackOnAiAnswer.service.ts
@@ -1,17 +1,18 @@
 import mongoose from "mongoose";
 
+import queueUserInterest from "../../../user/userInterest/queueUserInterest.service.js";
+
 import HttpError from "../../../../utils/http/httpError.util.js";
 import { makeJobId } from "../../../../utils/job/makeJobId.util.js";
-import queueUserInterest from "../../../../utils/question/queueUserInterest.util.js";
 
 import AiAnswer from "../../../../models/aiAnswer.model.js";
 import AiAnswerFeedback from "../../../../models/aiAnswerFeedback.model.js";
 import Question from "../../../../models/question.model.js";
 import QuestionVersion from "../../../../models/questionVersion.model.js";
 
-import contentFinalizeQueue from "../../../../queues/contentFinalize.queue.js";
-
 import { toPublicAiAnswerFeedback } from "../../question.response.js";
+
+import contentFinalizeQueue from "../../../../queues/contentFinalize.queue.js";
 
 const createFeedbackOnAiAnswerService = async (
   userId: string,
