+++ src/services/question/contentFinalize/contentFinalizeImage.service.ts
