+++ src/services/question/core/vote.service.ts
@@ -1,6 +1,7 @@
 import mongoose from "mongoose";
 
 import routeNotification from "../../notification/routeNotification.service.js";
+import queueUserInterest from "../../user/userInterest/queueUserInterest.service.js";
 
 import { getRedisCacheClient } from "../../../config/redis.config.js";
 
@@ -10,7 +11,6 @@ import {
   clearReplyCache,
 } from "../../../utils/cache/clearCache.util.js";
 import { makeJobId } from "../../../utils/job/makeJobId.util.js";
-import queueUserInterest from "../../../utils/question/queueUserInterest.util.js";
 
 import Question from "../../../models/question.model.js";
 import Answer from "../../../models/answer.model.js";
